Fix 1PPS output
In case the 1PPS output was enabled and the LTC was adjusted then also
the target clock was set in the future to adjust to continue to work.
The problem was that it was modifying also the nanosecond part, so then
the target clock became different that what was configured. Fix this by
not changing also the nanosecond part.