Disable parsing of PTP minor version
By default the PHY is checking the minor version inside the PTP header
and it expects to be a value of 0. This works fine with older ptp4l
versions as they are using a value of 0. With newer one it uses a value
of 1 to allign to 1588-2019 version.
To fix this, disable the PTP minor version check so that both versions
of ptp4l will work.