5.4: Change to receive the timestamp into the frame.
This is based on the upstream kernel which also has this change. This
improves the CPU performance because the CPU doesn't need to handle
anymore the RX interrupts to get the timestamp, as the timestamps are in
the frame.
This also fix an issue, for example if the frame is dropped by the MAC
HW then we would get the interrupt that the frame is timestamped but
then none one else will get out the timestmap from the fifo. So then
next time when timestamp a frame with the same seq number, we would
provide the old timestamp.